+++ R/ggadd.R
@@ -20,6 +20,10 @@ NULL
 #'  \code{\link{show_point_shapes}()}.
 #'@param size numeric value in [0-1] specifying point and line size.
 #'@param linetype line type.
+#'@param show.legend logical. Should this layer be included in the legends? NA,
+#'  the default, includes if any aesthetics are mapped. \code{FALSE} never includes,
+#'  and TRUE always includes. It can also be a named logical vector to finely
+#'  select the aesthetics to display.
 #'@param alpha numeric value specifying fill color transparency. Value should be
 #'  in [0, 1], where 0 is full transparency and 1 is no transparency.
 #'@param jitter a numeric value specifying the amount of jittering. Used only
@@ -44,7 +48,7 @@ NULL
 #'@export
 ggadd <- function(p, add = NULL, color = "black", fill = "white", group = 1,
                   width = 1, shape = 19, size = NULL, alpha = 1, jitter = 0.2,
-                  binwidth = NULL, dotsize = size, linetype = 1,
+                  binwidth = NULL, dotsize = size, linetype = 1, show.legend = NA,
                   error.plot = "pointrange", ci = 0.95,
                   data = NULL, position = position_dodge(0.8),
                   p_geom = ""
@@ -92,7 +96,8 @@ ggadd <- function(p, add = NULL, color = "black", fill = "white", group = 1,
   else if(ngrps > 1) jitter <- position_jitterdodge(jitter.width = .jitter, dodge.width = 0.8)
 
   common.opts <- opts <- list(data = data, color = color, fill = fill,
-                              size = size, position = position, alpha = alpha)
+                              size = size, position = position, alpha = alpha,
+                              show.legend = show.legend)
 
   if ("boxplot" %in% add) {
     if(.geom(p) == "violin" & missing(width)) width = 0.2
@@ -143,7 +148,8 @@ ggadd <- function(p, add = NULL, color = "black", fill = "white", group = 1,
       width <- 0.1
     else if(error.plot == "crossbar" & .geom(p) == "violin") width = 0.2
     p <- p %>% add_summary(errors, error.plot = error.plot, color = color, shape = shape,
-                  position = position, size = size, width = width, ci = ci, group = group, linetype = linetype)
+                  position = position, size = size, width = width, ci = ci, group = group,
+                  linetype = linetype, show.legend = show.legend)
   }
 
   p